Readability metrics (#5)
Added Flesch-Kincaid reading ease and grade level metrics and some related functionality: a sentence segmenter, a word syllable tokenizer, an optimized whitespace tokenizer (and additional regex for the same result with the RegexTokenizer), and some additional helpful functions on the Text API that were related.

Examples can now be found for some of the library (starting with readability examples) in the folder docs/examples. They are in the package main and can be run with a simple go run docs/examples/FILENAME.go.

Also started an ngrams package with a few small utility functions (they ended up not being used elsewhere at this time, but they are functional and could be useful, so they were not cut).